+++ omics_data_mgmt/users/urls.py
@@ -6,21 +6,11 @@ urlpatterns = [
     url(
         regex=r'^$',
         view=views.UserListView.as_view(),
-        name='list'
-    ),
-    url(
-        regex=r'^~redirect/$',
-        view=views.UserRedirectView.as_view(),
-        name='redirect'
+        name='user_list'
     ),
     url(
         regex=r'^(?P<username>[\w.@+-]+)/$',
         view=views.UserDetailView.as_view(),
-        name='detail'
-    ),
-    url(
-        regex=r'^~update/$',
-        view=views.UserUpdateView.as_view(),
-        name='update'
+        name='user_detail'
     ),
 ]