+++ projectroles/templatetags/projectroles_tags.py
@@ -148,7 +148,7 @@ def has_star(project, user):
 
 @register.simple_tag
 def get_help_highlight(user):
-    """Return classes to highlight navbar help link if user has recently 
+    """Return classes to highlight navbar help link if user has recently
     signed in"""
     if user.__class__.__name__ == 'User' and not user.is_anonymous:
         delta_days = (timezone.now() - user.date_joined).days